+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to the "L13 Diff" extension will be documented in this file.
 
 ## [0.14.0] - 2018-05-19
-- Added keyboard support for selection in list view.
+- Added keyboard support for selection in list view. See README.md
 
 ## [0.13.0] - 2018-05-12
 - Initial release