+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Change Log
 All notable changes to the "L13 Diff" extension will be documented in this file.
 
-## [0.14.0] - 2018-05-19
+## [0.14.0] - 2019-05-19
 - Added keyboard support for selection in list view. Please see README.md
 - Fixed a bug if an item was select in the menu with enter.
 - Fixed async loading bug if windows is running on a virtual machine.
 
-## [0.13.0] - 2018-05-12
+## [0.13.0] - 2019-05-12
 - Initial release